fitness_studio/requirements.txt
Django>=4.2
djangorestframework
pytz
